feat : index.component refactoring
