+++ src/app/components/widgets/medications-widget/medications-widget.component.ts
@@ -8,7 +8,7 @@ import {
   ApexLegend,ApexTitleSubtitle
 } from "ng-apexcharts";
 import { NgApexchartsModule } from "ng-apexcharts";
-import { Medication} from "../../../model/Medications"
+import { Medication } from '../../../model/medications';
 export type ChartOptions = {
   series: ApexAxisChartSeries | any;
   chart: ApexChart | any;
