--- src/app/model/Medication.ts
@@ -1,21 +1,0 @@
-export type Medication= {
-    Nom: string;
-    Dosage: string;
-    Forme: string;
-    Unit: string;
-    DCI: string;
-    Classe: string;
-    SousClasse: string;
-    Laboratoire: string;
-    AMM: string;
-    DateAMM: string;
-    Conditionnementprimaire: string;
-    Duréeconservation: number;
-    Indications: string;
-    StockDisponible: number;
-    StockMoyen: number;
-    StockMin:number;
-    StockAlerte: number;
-    StockSécurité: number;
-    StockMax: number;
-  };
