--- src/app/model/medication.ts
@@ -1,9 +1,0 @@
-export interface Medications {
-  [key: string]: any;
-  Nom: string;
-  Dosage: string;
-  Forme: string;
-  Unit: string;
-  DCI: string;
-  'Stock dispo': number;
-}
