+++ src/app/pages/dashboard/dashboard.component.html
@@ -155,7 +155,7 @@
                   <h3 class="text-white upgrade-info mb-0">
                     Medsecure - Doing now what <span class="fw-bold">patients</span> need next
                   </h3>
-                  <a href="#" class="btn btn-info upgrade-btn">Upgrade Account!</a>
+                  <a href="#" class="btn btn-info upgrade-btn">Create User Account</a>
                 </div>
               </div>
             </div>
