--- src/app/pages/page1/page1.component.html
@@ -1,3 +1,0 @@
-<br/>
-<p>page1 works!</p>
-<app-dialog></app-dialog>
