--- src/app/pages/page1/page1.component.spec.ts
@@ -1,23 +1,0 @@
-import { ComponentFixture, TestBed } from '@angular/core/testing';
-
-import { Page1Component } from './page1.component';
-
-describe('Page1Component', () => {
-  let component: Page1Component;
-  let fixture: ComponentFixture<Page1Component>;
-
-  beforeEach(async () => {
-    await TestBed.configureTestingModule({
-      imports: [Page1Component]
-    })
-    .compileComponents();
-    
-    fixture = TestBed.createComponent(Page1Component);
-    component = fixture.componentInstance;
-    fixture.detectChanges();
-  });
-
-  it('should create', () => {
-    expect(component).toBeTruthy();
-  });
-});
