+++ src/app/pages/widgets/widgets.component.ts
@@ -15,7 +15,7 @@ import { MultiSenseWidgetComponent} from '../../components/widgets/multisense-wi
 
 
 import {ArmoireStock } from "../../model/ArmoireStock"
-import { Medication} from "../../model/Medications"
+import { Medication } from './../../model/medications';
 import {Stockout} from "../../model/stockout"
 import { Presecription } from "../../model/Presecription"
 @Component({
