--- src/app/partials/_footer.html
@@ -1,9 +1,0 @@
-<footer class="footer">
-  <div class="d-sm-flex justify-content-center justify-content-sm-between">
-    <span class="text-muted text-center text-sm-left d-block d-sm-inline-block">Securing the medication circuit
-      within a healthcare facility by <a href="https://www.medsecure.com/" target="_blank">Medsecure
-      </a>App.</span>
-    <span class="float-none float-sm-end d-block mt-1 mt-sm-0 text-center">Copyright © 2024. All rights
-      reserved.</span>
-  </div>
-</footer>
