+++ src/app/partials/footer/footer.component.html
@@ -1,0 +1,12 @@
+<footer class="footer">
+    <div
+      class="d-sm-flex justify-content-center justify-content-sm-between"
+    >
+      <span
+        class="text-muted text-center text-sm-left d-block d-sm-inline-block"
+      >
+        Medsecure - Doing now what patients need next.
+        <span>Copyright © 2024. All rights reserved.</span>
+      </span>
+    </div>
+  </footer>
