+++ src/app/partials/settings-panel/settings-panel.component.ts
@@ -1,0 +1,11 @@
+import { Component } from '@angular/core';
+
+@Component({
+  selector: 'app-settings-panel',
+  standalone: true,
+  imports: [],
+  templateUrl: './settings-panel.component.html'
+})
+export class SettingsPanelComponent {
+
+}
